general cleanup and first attempt at reducing message noise.
